Pulled goecode code out of lib/index.js into it's own module
