+++ lib/index.js
@@ -1,27 +1,24 @@
-var request = require('request')
 var config = require('./config')
+var http = require('http');
+var low = require('lowdb')
+var moment = require('moment-timezone');
+var crypto = require('crypto');
+var logger = require('./logger')
+
 var routeNamesToRouteNumbers = require('./routename_to_routenumber');
 var stop_number_lookup = require('./stop_number_lookup')
-var turf = require('turf')
-var all_stops = require('../gtfs/geojson/stops.json');
 var exceptions = require('../gtfs/geojson/exceptions.json');
-var moment = require('moment-timezone');
-var low = require('lowdb')
 var comments = low('./comments.json', { storage: require('lowdb/lib/file-async') });
-var http = require('http');
-var nodemailer = require('nodemailer');
-var transporter = nodemailer.createTransport({
+
+var transporter = require('nodemailer').createTransport({
     service: 'gmail',
     auth: {
         user: config.GMAIL_USERNAME,
         pass: config.GMAIL_PASSWORD
     }
-}, {
+    }, {
     from: config.FEEDBACK_EMAIL_FROM
 });
-var crypto = require('crypto');
-var logger = require('./logger')
-
 
 /*
 * @function getStopFromStopNumber
@@ -37,51 +34,10 @@ function getStopFromStopNumber(stopId) {
         return Promise.reject(err)
     }
     return requestBusData(busTrackerId)
-    .then((muniBusData)=>{
-        return {data: parseBusData(muniBusData.data, stopId), muniTime:muniBusData.asyncTime};
-    })
+    .then(muniBusData => ({data: parseBusData(muniBusData.data, stopId), muniTime:muniBusData.asyncTime}))
     .catch((err) => Promise.reject(new Error('Sorry, Bustracker is down')))
 }
 
-/*
-* @function getStopsFromAddress
-* @param address - almost any text you think Google can figure out
-* @returns an Promise which is: {data: {stops:[found-stops],geocodedAddress:theaddress}, geocodeTime: timing-of-Google-request  }
-*
-* asyncTime is how long the goecoder took
-*/
-function NotFoundError(message) {
-    this.name = "Address not found"
-    this.message = message
-    this.stack = Error().stack
-    this.type = 'NOT_FOUND'
-}
-NotFoundError.prototype = new Error;
-
-function getStopsFromAddress(address){
-    return exports.getGeocodedAddress(address)
-    .then((returnObj) => {
-        var geocodedPlace = returnObj.data
-        if (!geocodedPlace){
-            return Promise.reject(new NotFoundError(`Searched for "${address}"`))
-        }
-        var address = geocodedPlace.formatted_address
-        var lat = geocodedPlace.location.lat;
-        var lon = geocodedPlace.location.lng;
-
-        // Stops can be empty. If no stops are found within max distance
-        // this returns an empty array for the data.stops so routes can decide what to do.
-        var stops = findNearestStops(lat, lon);
-
-        return {data: {stops: stops, geocodedAddress:address} , geocodeTime: returnObj.asyncTime};
-    })
-    .catch(err => {
-        if (err.type == 'NOT_FOUND') return Promise.reject(err)
-        // A real error from the goecoder must have occured to make it here.
-        logger.error(err)
-        return Promise.reject(new NotFoundError(`Searched for "${address}"`)) // This lies to the user if there is a geocoder error.
-    })
-}
 
 /**
 * @function requestBusData
@@ -92,12 +48,12 @@ function getStopsFromAddress(address){
 function requestBusData(busTrackerId) {
     return new Promise((resolve, reject) => {
         var asyncTime =  Date.now()
-        var request = http.get(config.MUNI_URL + busTrackerId, (response) => {
+        var request = http.get(config.MUNI_URL + busTrackerId, response => {
             if (response.statusCode < 200 || response.statusCode > 299) {
                 reject(new Error('Failed to load bustracker, status code: ' + response.statusCode));
             }
             const body = [];
-            response.on('data', (chunk) => body.push(chunk));
+            response.on('data', chunk => body.push(chunk));
             response.on('end', () => {
                 resolve({
                     data: body.join(''),
@@ -140,7 +96,7 @@ function parseBusData(body, stopId) {
     }
 
     var currentStop = null;
-    stopsAndTimes.forEach(function(stopOrTime) {
+    stopsAndTimes.forEach(stopOrTime => {
         if (stopOrTime === 'Done') {
             currentStop.times.push('Out of Service')
         }
@@ -173,80 +129,6 @@ function prependRouteNumberToRouteName(routeName) {
     return {routeName: routeName};
 }
 
-function stopLatLong(stopid) {
-    for (var i=0; i < all_stops.features.length; i++) {
-        if (all_stops.features[i].properties.stop_id.match(/\d+/g)[0] == stopid) {
-            return all_stops.features[i].geometry.coordinates[1] + "," + all_stops.features[i].geometry.coordinates[0];
-        }
-    }
-    return "";
-}
-
-/**
-* @function findNearestStops
-* @returns object {route: stopId: distance: ll:}
-*
-*/
-function findNearestStops(lat, lon) {
-    var point = turf.point([lon, lat]);
-    var buffer = turf.buffer(point, config.NEAREST_BUFFER, 'miles');
-    var nearest_stops = turf.within(all_stops, buffer);
-    var out = nearest_stops.features.map(function(stop){
-        var stopId = stop.properties.stop_id.match(/\d+/g)[0];
-        return { route: stop.properties.name,
-                 stopId: stopId,
-                 distance: turf.distance(point, stop, "miles"),
-                 ll: stopLatLong(stopId)
-        }
-    });
-    // returns empty if none found nearby
-    out.sort(function(a, b) {
-        return (a.distance - b.distance)
-    });
-    if (out.length > config.NEAREST_MAX) out = out.slice(0,config.NEAREST_MAX);
-    return out;
-}
-
-function getGeocodedAddress(address) {
-    // TODO: what happens when we reach the free limit?
-    var GEOCODE_URL_BASE = "https://maps.googleapis.com/maps/api/geocode/json?"
-    var CITY = config.GOOGLE_GEOCODE_LOCATION
-    var COUNTRY = "US"
-    var timer = Date.now();
-    return new Promise((resolve, reject) => {
-        var querry = encodeURIComponent(address)
-        request(`${GEOCODE_URL_BASE}address=${querry}&components=country:${COUNTRY}|administrative_area:${CITY}&key=${process.env.GOOGLE_MAPS_KEY}`, function(error, response, body){
-            if (!error && response.statusCode == 200) {
-                var geocodeData = JSON.parse(body)
-                if (geocodeData.status != "OK") {
-                    return reject(new Error(`Geocoder Error: ${geocodeData.status}`))
-                }
-                var acceptable_types = [
-                    'route',
-                    'street_address',
-                    'intersection',
-                    'transit_station',
-                    'point_of_interest',
-                    'establishment',
-                    'train_station',
-                    'bus_station',
-                    'neighborhood'
-                ]
-                var result = geocodeData.results[0]
-
-                var data = (result.types && acceptable_types.some(el => result.types.includes(el)))
-                    ? {location:result.geometry.location, formatted_address: result.formatted_address}
-                    : null
-                resolve({data: data, asyncTime: Date.now()-timer})
-
-            } else {
-                reject(error)
-            }
-        })
-
-    })
-}
-
 function processFeedback(feedback, req) {
     var response_hash = crypto.randomBytes(20).toString('hex');
 
@@ -277,19 +159,7 @@ function processFeedback(feedback, req) {
 function serviceExceptions() {
     // Check for holiday exceptions
     var dateTz = moment.tz(new Date(), config.TIMEZONE).format("YYYYMMDD");
-    var noService = false;
-    var replacementService = false;
-    for(var i=0; (i < exceptions.exceptions.length) && !replacementService; i++) {
-        if (exceptions.exceptions[i].date == dateTz) {
-            if (exceptions.exceptions[i].exception_type == 1) {
-                replacementService = true;
-                noService = false;
-            } else {
-                noService = true;
-            }
-        }
-    }
-    return noService;
+    return exceptions.exceptions.some(exception => exception.date == dateTz && exception.exception_type == 2 )
 }
 
 /*
@@ -323,15 +193,12 @@ function verifyFBRequestSignature(req, res, buf) {
     }
 }
 
-module.exports.getStopsFromAddress = getStopsFromAddress;
 module.exports.getStopFromStopNumber = getStopFromStopNumber;
 module.exports.serviceExceptions = serviceExceptions;
-module.exports.findNearestStops = findNearestStops;
 module.exports.processFeedback = processFeedback;
 module.exports.verifyFBRequestSignature = verifyFBRequestSignature;
 
 // export this for unit testing purposes
-module.exports.getGeocodedAddress = getGeocodedAddress
 
 /*
 module.exports.getStopFromBusTrackerId = getStopFromBusTrackerId;
