+++ package.json
@@ -38,8 +38,9 @@
     "rollbar": "^0.6.2",
     "run-middleware": "^0.6.6",
     "serve-favicon": "~2.2.0",
-    "turf": "~1.4.0",
-    "turf-nearest": "~1.0.2",
+    "@turf/circle": "^4.3.0",
+    "@turf/distance": "^4.3.0",
+    "@turf/within": "^4.3.0",
     "twilio": "^3.3.0-edge",
     "universal-analytics": "^0.4.8",
     "watson-developer-cloud": "^2.9.1",
